+++ lib/streamlit/web/bootstrap.py
@@ -30,7 +30,7 @@ from streamlit import version
 from streamlit.config import CONFIG_FILENAMES
 from streamlit.git_util import GitRepo, MIN_GIT_VERSION
 from streamlit.logger import get_logger
-from streamlit.secrets import SECRETS_FILE_LOC
+from streamlit.runtime.secrets import SECRETS_FILE_LOC
 from streamlit.source_util import invalidate_pages_cache
 from streamlit.watcher import report_watchdog_availability, watch_dir, watch_file
 from streamlit.web.server import Server, server_address_is_unix_socket
